+++ packages/chain-adapters/src/utxo/UTXOBaseAdapter.ts
@@ -40,7 +40,7 @@ export type UTXOChainTypes = KnownChainIds.BitcoinMainnet // to be extended in t
 export interface ChainAdapterArgs {
   providers: {
     http: unchained.bitcoin.V1Api
-    ws: unchained.ws.Client<unchained.Tx>
+    ws: unchained.ws.Client<unchained.bitcoin.BitcoinTx>
   }
   coinName: string
   chainId?: ChainId
@@ -56,10 +56,11 @@ export abstract class UTXOBaseAdapter<T extends UTXOChainTypes> implements IChai
   protected chainId: ChainId
   protected assetId: AssetId
   protected coinName: string
+  protected accountAddresses: Record<string, Array<string>> = {}
   protected readonly supportedChainIds: ChainId[]
   protected readonly providers: {
     http: unchained.bitcoin.V1Api
-    ws: unchained.ws.Client<unchained.Tx>
+    ws: unchained.ws.Client<unchained.bitcoin.BitcoinTx>
   }
 
   protected constructor(args: ChainAdapterArgs) {
@@ -109,6 +110,11 @@ export abstract class UTXOBaseAdapter<T extends UTXOChainTypes> implements IChai
       const { data } = await this.providers.http.getAccount({ pubkey })
 
       const balance = bnOrZero(data.balance).plus(bnOrZero(data.unconfirmedBalance))
+
+      // cache addresses for getTxHistory to use without needing to make extra requests
+      this.accountAddresses[data.pubkey] = data.addresses?.map((address) => address.pubkey) ?? [
+        data.pubkey
+      ]
 
       return {
         balance: balance.toString(),