+++ packages/swapper/src/swappers/zrx/zrxBuildTrade/zrxBuildTrade.test.ts
@@ -43,7 +43,8 @@ const setup = () => {
   const unchainedUrls = {
     [KnownChainIds.EthereumMainnet]: {
       httpUrl: 'http://localhost:31300',
-      wsUrl: 'ws://localhost:31300'
+      wsUrl: 'ws://localhost:31300',
+      rpcUrl: 'http://localhost:1000'
     }
   }
   const ethNodeUrl = 'http://localhost:1000'